+++ packages/stream_video/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+🐞 Fixed
+* Fixed incoming call timeout handling.
+  - Use `streamVideo.observeCoreRingingEventsForBackground()` instead of `streamVideo.observeCallDeclinedRingingEvent()` in `firebaseMessagingBackgroundHandler` to support all necessary events.
+
 * Improved SFU stats implementation.
 
 ## 1.0.1